learn-android/replsdk/src/main/java/com/dbs/replsdk/model/RequestContext.java
package com.dbs.replsdk.model;

public class RequestContext {

    private Device device;
    private Platform platform;
    private User user;
    private String api_version;

    public Device getDevice() {
        return device;
    }

    public void setDevice(Device device) {
        this.device = device;
    }

    public Platform getPlatform() {
        return platform;
    }

    public void setPlatform(Platform platform) {
        this.platform = platform;
    }

    public User getUser() {
        return user;
    }

    public void setUser(User user) {
        this.user = user;
    }

    public String getApi_version() {
        return api_version;
    }

    public void setApi_version(String api_version) {
        this.api_version = api_version;
    }


}
